+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/model/BlockBatch.kt
@@ -1,0 +1,9 @@
+package cash.z.ecc.android.sdk.model
+
+import cash.z.ecc.android.sdk.internal.model.JniBlockMeta
+
+data class BlockBatch(
+    val index: Long,
+    val range: ClosedRange<BlockHeight>,
+    var blocks: List<JniBlockMeta>? = null
+)